Refactor: 공통 컴포넌트 atomic 패턴에 맞게 재조정
