+++ src/components/atoms/index.ts
@@ -1,3 +1,4 @@
+export { default as Button } from './Button';
 export { default as Input } from './Input';
-export { default as Txt } from './Text';
 export { default as ShadowBox } from './ShadowBox';
+export { default as Txt } from './Text';
