+++ src/components/molecules/index.ts
@@ -1,4 +1,1 @@
-export { default as Button } from './Button';
-export { default as Header } from './Header';
-export { default as FixedBottomButton } from '../organisms/FixedBottomButton';
 export { default as BottomTabNavBtn } from './BottomTabNavBtn';
