+++ src/components/organisms/index.ts
@@ -1,3 +1,4 @@
-export { default as Modal } from './Modal';
-export { default as Header } from './Header';
 export { default as BottomTabNav } from './BottomTabNav';
+export { default as FixedBottomButton } from './FixedBottomButton';
+export { default as Header } from './Header';
+export { default as Modal } from './Modal';
